requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
flake8==7.2.0
idna==3.10
mccabe==0.7.0
pycodestyle==2.13.0
pyflakes==3.3.2
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
requests==2.32.3
urllib3==2.4.0
